Pig Claim script
Note: The source input files are hardcoded because, first I tried using
the properties file to set the input paths, and when i was running pig
from my machine seems to worked fine, but pig complained about some LZO
compression, so i started to use my VM with a Cloudera and within this
env, i faced some .properties file not found, so in order to complete
this I hard code the paths. I will be reviewing this later.